Works fine without Int8 convertion.

+++ src/geoip2/maxminddb.cr
@@ -16,8 +16,8 @@ module GeoIP2
     fun lookup_string = MMDB_lookup_string(
                                            mmdb : MMDB*,
                                            ipstr : Char*,
-                                           gai_error : Int8*,
-                                           mmdb_error : Int8*) : LookupResultS
+                                           gai_error : Int*,
+                                           mmdb_error : Int*) : LookupResultS
 
     fun lookup_sockaddr = MMDB_lookup_sockaddr(
                                                mmdb : MMDB*,
